aureon-cortex/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
supabase>=2.3.0
# AI Core
langchain>=0.1.0
google-generativeai>=0.4.0
# Utilities
loguru>=0.7.2
python-telegram-bot
python-multipart>=0.0.9
email-validator>=2.1.0
google-genai
pydub
edge-tts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic-ai
nest_asyncio
prometheus-fastapi-instrumentator
orjson
uvloop; sys_platform != 'win32'
httptools; sys_platform != 'win32'
